lab5/components/data_ingestor/requirements.txt
google-cloud-storage
panda
pathlib